requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
marshmallow==3.20.1
